Componentes en directorios de usuarios

+++ app/Http/Controllers/Business/BusinessController.php
@@ -15,53 +15,65 @@ class BusinessController extends Controller {
 	}
 
 	public function get(){
-		$data = [
-					'status'=>true,
-					'business'=>Business::get()
-			    ];
-		return response()->json($data);
+		if (auth()->user()->roles->pluck('name')[0] == 'super.admin') {
+			$data = [
+						'status'=>true,
+						'business'=>Business::get()
+				    ];
+			return response()->json($data);
+		}
+		return $this->unauthorizedAccess();
 	}
 
 	public function store(Request $request){
-		$business = new Business($request->business);
-		$business->save();
-
-		$data = [
-					'status' => true,
-					'business' => $business
-				];
-
-		return response()->json($data);
+		if (auth()->user()->roles->pluck('name')[0] == 'super.admin') {
+			$business = new Business($request->business);
+			$business->save();
+			$data = [
+						'status' => true,
+						'business' => $business
+					];
+			return response()->json($data);
+		}
+		return $this->unauthorizedAccess();
 	}
 
 	public function update(Business $business, Request $request){
-		$business->update($request->business);
-
-		$data = [
-					'status' => true,
-					'business' => $request->business
-				];
-
-		return response()->json($data);
+		if (auth()->user()->roles->pluck('name')[0] == 'super.admin') {
+			$business->update($request->business);
+			$data = [
+						'status' => true,
+						'business' => $request->business
+					];
+			return response()->json($data);
+		}
+		return $this->unauthorizedAccess();
 	}
 
 	public function delete(Business $business){
-		$business->delete();
-        return response()->json(['status' => true]);
+		if (auth()->user()->roles->pluck('name')[0] == 'super.admin') {
+			$business->delete();
+	        return response()->json(['status' => true]);
+    	}
+    	return $this->unauthorizedAccess();
 	}
 
 	public function getCollectors(Business $business){
-		
-		$collectors = User::with('personal_information')->where('business_id','=',$business->id)
-														->role('cobrador')
-														->get();
-		
+		if (auth()->user()->roles->pluck('name')[0] == 'super.admin' || auth()->user()->roles->pluck('name')[0] == 'admin') {
+			$collectors = User::with('personal_information')->where('business_id','=',$business->id)
+															->role('cobrador')
+															->get();
 
-		/*$collector_aux = PersonalInformation::whereHas('user', function($query) use ($business){
-			$query->where('business_id','=',$business->id);
-		})->get();*/
-
-        return response()->json(['status'=>true, 'business'=>$business, 'collectors'=>$collectors]);
+	        return response()->json(['status'=>true, 'business'=>$business, 'collectors'=>$collectors]);
+	    }
+    	return $this->unauthorizedAccess();
 	}
 
+	public function unauthorizedAccess(){
+		$data = [
+					'status' => false,
+					'codigo' => 401
+				];
+		return response()->json($data);
+	}
 }
